Rotate labels if they overlap, hide legend if it overlaps with chart. Hide xAxis if labels are rotated and still overlap
add callback on renderEnd

ResizableChartDemo emit event on resizeStop
