+++ packages/superset-ui-demo/storybook/stories/plugins/legacy-preset-chart-nvd3/Area/data.ts
@@ -1125,4 +1125,229 @@ export default [
       },
     ],
   },
+  {
+    key: ['Latin America & Caribbean'],
+    values: [
+      {
+        x: -315619200000.0,
+        y: 220564224.0,
+      },
+      {
+        x: -283996800000.0,
+        y: 226764342.0,
+      },
+      {
+        x: -252460800000.0,
+        y: 233183206.0,
+      },
+      {
+        x: -220924800000.0,
+        y: 239771182.0,
+      },
+      {
+        x: -189388800000.0,
+        y: 246458356.0,
+      },
+      {
+        x: -157766400000.0,
+        y: 253195267.0,
+      },
+      {
+        x: -126230400000.0,
+        y: 259965218.0,
+      },
+      {
+        x: -94694400000.0,
+        y: 266776414.0,
+      },
+      {
+        x: -63158400000.0,
+        y: 273654630.0,
+      },
+      {
+        x: -31536000000.0,
+        y: 280641049.0,
+      },
+      {
+        x: 0.0,
+        y: 287763515.0,
+      },
+      {
+        x: 31536000000.0,
+        y: 295026304.0,
+      },
+      {
+        x: 63072000000.0,
+        y: 302408883.0,
+      },
+      {
+        x: 94694400000.0,
+        y: 309902169.0,
+      },
+      {
+        x: 126230400000.0,
+        y: 317479496.0,
+      },
+      {
+        x: 157766400000.0,
+        y: 325120067.0,
+      },
+      {
+        x: 189302400000.0,
+        y: 332817916.0,
+      },
+      {
+        x: 220924800000.0,
+        y: 340569396.0,
+      },
+      {
+        x: 252460800000.0,
+        y: 348391181.0,
+      },
+      {
+        x: 283996800000.0,
+        y: 356288443.0,
+      },
+      {
+        x: 315532800000.0,
+        y: 364270961.0,
+      },
+      {
+        x: 347155200000.0,
+        y: 372330102.0,
+      },
+      {
+        x: 378691200000.0,
+        y: 380466998.0,
+      },
+      {
+        x: 410227200000.0,
+        y: 388654061.0,
+      },
+      {
+        x: 441763200000.0,
+        y: 396869481.0,
+      },
+      {
+        x: 473385600000.0,
+        y: 405083115.0,
+      },
+      {
+        x: 504921600000.0,
+        y: 413292690.0,
+      },
+      {
+        x: 536457600000.0,
+        y: 421490233.0,
+      },
+      {
+        x: 567993600000.0,
+        y: 429668211.0,
+      },
+      {
+        x: 599616000000.0,
+        y: 437843614.0,
+      },
+      {
+        x: 631152000000.0,
+        y: 445998222.0,
+      },
+      {
+        x: 662688000000.0,
+        y: 454117634.0,
+      },
+      {
+        x: 694224000000.0,
+        y: 462201058.0,
+      },
+      {
+        x: 725846400000.0,
+        y: 470263697.0,
+      },
+      {
+        x: 757382400000.0,
+        y: 478310786.0,
+      },
+      {
+        x: 788918400000.0,
+        y: 486343677.0,
+      },
+      {
+        x: 820454400000.0,
+        y: 494384205.0,
+      },
+      {
+        x: 852076800000.0,
+        y: 502390020.0,
+      },
+      {
+        x: 883612800000.0,
+        y: 510356845.0,
+      },
+      {
+        x: 915148800000.0,
+        y: 518188225.0,
+      },
+      {
+        x: 946684800000.0,
+        y: 525886558.0,
+      },
+      {
+        x: 978307200000.0,
+        y: 533449671.0,
+      },
+      {
+        x: 1009843200000.0,
+        y: 540884684.0,
+      },
+      {
+        x: 1041379200000.0,
+        y: 548225528.0,
+      },
+      {
+        x: 1072915200000.0,
+        y: 555515431.0,
+      },
+      {
+        x: 1104537600000.0,
+        y: 562783235.0,
+      },
+      {
+        x: 1136073600000.0,
+        y: 570029991.0,
+      },
+      {
+        x: 1167609600000.0,
+        y: 577248307.0,
+      },
+      {
+        x: 1199145600000.0,
+        y: 584435842.0,
+      },
+      {
+        x: 1230768000000.0,
+        y: 591577623.0,
+      },
+      {
+        x: 1262304000000.0,
+        y: 598662941.0,
+      },
+      {
+        x: 1293840000000.0,
+        y: 605674766.0,
+      },
+      {
+        x: 1325376000000.0,
+        y: 612617659.0,
+      },
+      {
+        x: 1356998400000.0,
+        y: 619487273.0,
+      },
+      {
+        x: 1388534400000.0,
+        y: 626270167.0,
+      },
+    ],
+  },
 ];
