+++ packages/superset-ui-demo/storybook/stories/plugins/legacy-preset-chart-nvd3/Bar/stories/basic.tsx
@@ -2,35 +2,40 @@ import React from 'react';
 import { SuperChart } from '@superset-ui/core';
 import dummyDatasource from '../../../../../shared/dummyDatasource';
 import data from '../data';
+import ResizableChartDemo from '../../../../../shared/components/ResizableChartDemo';
 
 export const basic = () => (
-  <SuperChart
-    chartType="bar"
-    width={400}
-    height={400}
-    datasource={dummyDatasource}
-    queriesData={[{ data }]}
-    formData={{
-      bottomMargin: 'auto',
-      colorScheme: 'd3Category10',
-      contribution: false,
-      groupby: ['region'],
-      lineInterpolation: 'linear',
-      metrics: ['sum__SP_POP_TOTL'],
-      richTooltip: true,
-      showBarValue: false,
-      showBrush: 'auto',
-      showControls: false,
-      showLegend: true,
-      stackedStyle: 'stack',
-      vizType: 'bar',
-      xAxisFormat: '%Y',
-      xAxisLabel: '',
-      xAxisShowminmax: false,
-      xTicksLayout: 'auto',
-      yAxisBounds: [null, null],
-      yAxisFormat: '.3s',
-      yLogScale: false,
-    }}
-  />
+  <ResizableChartDemo>
+    {({ width, height }) => (
+      <SuperChart
+        chartType="bar"
+        width={width}
+        height={height}
+        datasource={dummyDatasource}
+        queriesData={[{ data }]}
+        formData={{
+          bottomMargin: 'auto',
+          colorScheme: 'd3Category10',
+          contribution: false,
+          groupby: ['region'],
+          lineInterpolation: 'linear',
+          metrics: ['sum__SP_POP_TOTL'],
+          richTooltip: true,
+          showBarValue: false,
+          showBrush: 'auto',
+          showControls: false,
+          showLegend: true,
+          stackedStyle: 'stack',
+          vizType: 'bar',
+          xAxisFormat: '%Y-%m-%d %H:%M:%S',
+          xAxisLabel: '',
+          xAxisShowminmax: false,
+          xTicksLayout: 'auto',
+          yAxisBounds: [null, null],
+          yAxisFormat: '.3s',
+          yLogScale: false,
+        }}
+      />
+    )}
+  </ResizableChartDemo>
 );
