+++ packages/superset-ui-demo/storybook/stories/plugins/legacy-preset-chart-nvd3/BoxPlot/stories/basic.tsx
@@ -2,18 +2,23 @@ import React from 'react';
 import { SuperChart } from '@superset-ui/core';
 import dummyDatasource from '../../../../../shared/dummyDatasource';
 import data from '../data';
+import ResizableChartDemo from '../../../../../shared/components/ResizableChartDemo';
 
 export const basic = () => (
-  <SuperChart
-    chartType="box-plot"
-    width={400}
-    height={400}
-    datasource={dummyDatasource}
-    queriesData={[{ data }]}
-    formData={{
-      colorScheme: 'd3Category10',
-      vizType: 'box_plot',
-      whiskerOptions: 'Min/max (no outliers)',
-    }}
-  />
+  <ResizableChartDemo>
+    {({ width, height }) => (
+      <SuperChart
+        chartType="box-plot"
+        width={width}
+        height={height}
+        datasource={dummyDatasource}
+        queriesData={[{ data }]}
+        formData={{
+          colorScheme: 'd3Category10',
+          vizType: 'box_plot',
+          whiskerOptions: 'Min/max (no outliers)',
+        }}
+      />
+    )}
+  </ResizableChartDemo>
 );
