+++ packages/superset-ui-demo/storybook/stories/plugins/legacy-preset-chart-nvd3/Bullet/stories/basic.tsx
@@ -2,22 +2,27 @@ import React from 'react';
 import { SuperChart } from '@superset-ui/core';
 import dummyDatasource from '../../../../../shared/dummyDatasource';
 import data from '../data';
+import ResizableChartDemo from '../../../../../shared/components/ResizableChartDemo';
 
 export const basic = () => (
-  <SuperChart
-    chartType="bullet"
-    width={400}
-    height={400}
-    datasource={dummyDatasource}
-    queriesData={[{ data }]}
-    formData={{
-      markerLabels: '',
-      markerLineLabels: '',
-      markerLines: '',
-      markers: '',
-      rangeLabels: '',
-      ranges: '',
-      vizType: 'bullet',
-    }}
-  />
+  <ResizableChartDemo>
+    {({ width, height }) => (
+      <SuperChart
+        chartType="bullet"
+        width={width}
+        height={height}
+        datasource={dummyDatasource}
+        queriesData={[{ data }]}
+        formData={{
+          markerLabels: '',
+          markerLineLabels: '',
+          markerLines: '',
+          markers: '',
+          rangeLabels: '',
+          ranges: '',
+          vizType: 'bullet',
+        }}
+      />
+    )}
+  </ResizableChartDemo>
 );
