+++ packages/superset-ui-demo/storybook/stories/plugins/legacy-preset-chart-nvd3/DistBar/stories/basic.tsx
@@ -2,29 +2,34 @@ import React from 'react';
 import { SuperChart } from '@superset-ui/core';
 import dummyDatasource from '../../../../../shared/dummyDatasource';
 import data from '../data';
+import ResizableChartDemo from '../../../../../shared/components/ResizableChartDemo';
 
 export const basic = () => (
-  <SuperChart
-    chartType="dist-bar"
-    width={400}
-    height={400}
-    datasource={dummyDatasource}
-    queriesData={[{ data }]}
-    formData={{
-      barstacked: false,
-      bottomMargin: 'auto',
-      colorScheme: 'd3Category10',
-      contribution: false,
-      orderBars: false,
-      reduceXTicks: false,
-      showBarValue: false,
-      showControls: false,
-      showLegend: true,
-      vizType: 'dist_bar',
-      xAxisLabel: 'ddd',
-      xTicksLayout: 'auto',
-      yAxisFormat: '.3s',
-      yAxisLabel: 'ddd',
-    }}
-  />
+  <ResizableChartDemo>
+    {({ width, height }) => (
+      <SuperChart
+        chartType="dist-bar"
+        width={width}
+        height={height}
+        datasource={dummyDatasource}
+        queriesData={[{ data }]}
+        formData={{
+          barstacked: false,
+          bottomMargin: 'auto',
+          colorScheme: 'd3Category10',
+          contribution: false,
+          orderBars: false,
+          reduceXTicks: false,
+          showBarValue: false,
+          showControls: false,
+          showLegend: true,
+          vizType: 'dist_bar',
+          xAxisLabel: 'ddd',
+          xTicksLayout: 'auto',
+          yAxisFormat: '.3s',
+          yAxisLabel: 'ddd',
+        }}
+      />
+    )}
+  </ResizableChartDemo>
 );
