+++ packages/superset-ui-demo/storybook/stories/plugins/legacy-preset-chart-nvd3/Line/stories/logScale.tsx
@@ -2,20 +2,25 @@ import React from 'react';
 import { SuperChart } from '@superset-ui/core';
 import dummyDatasource from '../../../../../shared/dummyDatasource';
 import data from '../data';
+import ResizableChartDemo from '../../../../../shared/components/ResizableChartDemo';
 
 export const logScale = () => (
-  <SuperChart
-    chartType="line"
-    width={400}
-    height={400}
-    datasource={dummyDatasource}
-    queriesData={[{ data }]}
-    formData={{
-      richTooltip: true,
-      vizType: 'line',
-      yAxisBounds: [1, 60000],
-      yAxisFormat: ',d',
-      yLogScale: true,
-    }}
-  />
+  <ResizableChartDemo>
+    {({ width, height }) => (
+      <SuperChart
+        chartType="line"
+        width={width}
+        height={height}
+        datasource={dummyDatasource}
+        queriesData={[{ data }]}
+        formData={{
+          richTooltip: true,
+          vizType: 'line',
+          yAxisBounds: [1, 60000],
+          yAxisFormat: ',d',
+          yLogScale: true,
+        }}
+      />
+    )}
+  </ResizableChartDemo>
 );
