+++ plugins/plugin-chart-table/src/transformProps.ts
@@ -24,7 +24,6 @@ import {
   getNumberFormatter,
   getTimeFormatter,
   getTimeFormatterForGranularity,
-  LOCAL_PREFIX,
   NumberFormats,
   QueryMode,
   smartDateFormatter,
@@ -117,7 +116,7 @@ const processColumns = memoizeOne(function processColumns(props: TableChartProps
             formatter = getTimeFormatterForGranularity(granularity);
           } else if (format) {
             // other columns respect the column-specific format
-            formatter = getTimeFormatter(LOCAL_PREFIX + format);
+            formatter = getTimeFormatter(format);
           } else if (isNumeric(key, records)) {
             // if column is numeric values, it is considered a timestamp64
             formatter = getTimeFormatter(DATABASE_DATETIME);
@@ -126,7 +125,7 @@ const processColumns = memoizeOne(function processColumns(props: TableChartProps
             formatter = String;
           }
         } else if (timeFormat) {
-          formatter = getTimeFormatter(LOCAL_PREFIX + timeFormat);
+          formatter = getTimeFormatter(timeFormat);
         }
       } else if (isMetric) {
         formatter = getNumberFormatter(format);
